support event recorder and replay

+++ src/tkc/event.h
@@ -99,11 +99,11 @@ typedef struct _event_t {
    */
   uint32_t type;
   /**
-   * @property {int32_t} time
+   * @property {uint64_t} time
    * @annotation ["readable", "scriptable"]
    * 事件发生的时间。
    */
-  uint32_t time;
+  uint64_t time;
   /**
    * @property {void*} target
    * @annotation ["readable", "scriptable"]
